backend/requirements.txt
fastapi
langchain
openai
pinecone-client	
python-dotenv
httpx	
uvicorn